#33 wip split currency types
