+++ src/features/currency/api/client.ts
@@ -1,7 +1,7 @@
 import { API_BASE_URL } from '../../../config/env';
 import { Cached } from '../../../lib/cached';
 import type { Results } from '../../../types/results';
-import { PIVOT_CURRENCY_CODE } from '../constants';
+import { PIVOT_CURRENCY_CODE } from '../config/codes';
 import type {
   Currency,
   DateRate,
